Add AbstractHttpRSocketFunction class.
