+++ rsocket-routing-http-bridge/src/main/java/io/rsocket/routing/http/bridge/config/RSocketHttpBridgeAutoConfiguration.java
@@ -3,7 +3,8 @@ package io.rsocket.routing.http.bridge.config;
 import java.util.function.Function;
 
 import io.rsocket.routing.client.spring.RoutingRSocketRequester;
-import io.rsocket.routing.http.bridge.core.HttpRSocketFunction;
+import io.rsocket.routing.http.bridge.core.RequestResponseFunction;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,9 +24,29 @@ public class RSocketHttpBridgeAutoConfiguration {
 	@Autowired
 	RoutingRSocketRequester requester;
 
+	// Stay with four different endpoints or switch to some other way of differentiating?
+
 	@Bean
-	public Function<Mono<Message<String>>, Mono<Message<String>>> function() {
-		return new HttpRSocketFunction(requester);
+	public Function<Mono<Message<String>>, Mono<Message<String>>> rr() {
+		return new RequestResponseFunction(requester);
+	}
+
+	@Bean
+	public Function<Flux<Message<String>>, Flux<Message<String>>> rc() {
+		// TODO
+		return messageFlux -> messageFlux.map(message -> message);
+	}
+
+	@Bean
+	public Function<Mono<Message<String>>, Flux<Message<String>>> rs() {
+		// TODO
+		return messageMono -> messageMono.map(message -> message).flux();
+	}
+
+	@Bean
+	public Function<Mono<Message<String>>, Mono<Void>> ff() {
+		// TODO
+		return messageMono -> Mono.empty();
 	}
 
 }
