+++ rsocket-routing-http-bridge/src/main/java/io/rsocket/routing/http/bridge/core/AbstractHttpRSocketFunction.java
@@ -1,0 +1,33 @@
+package io.rsocket.routing.http.bridge.core;
+
+import java.time.Duration;
+import java.util.function.Function;
+
+import org.apache.commons.logging.Log;
+
+import static org.apache.commons.logging.LogFactory.getLog;
+
+/**
+ * @author Olga Maciaszek-Sharma
+ */
+public abstract class AbstractHttpRSocketFunction<I, O> implements Function<I, O> {
+
+	protected final Log LOG = getLog(getClass());
+
+	// TODO: get from properties; initialise in constructor
+	protected final Duration timeout = Duration.ofSeconds(30);
+
+	void logTimeout(String address, String route) {
+		if (LOG.isDebugEnabled()) {
+			LOG.debug(String
+					.format("Timeout occurred while retrieving RSocket response from address: %s, route: %s. Response was not retrieved within %s", address, route, timeout));
+		}
+	}
+
+	void logException(Throwable error, String address, String route) {
+		if (LOG.isErrorEnabled())
+			LOG.error(String
+					.format("Exception occurred while retrieving RSocket response from address: %s, route: %s", address, route), error);
+	}
+
+}
